requirements.txt
tkinter
python-dotenv
google-generativeai 